jason: camera: Clear xiaomi reversed field before processing 3A metadata
 * (Perhaps) magically fixed failure in some usecase.
 * Correct the data type.

+++ camera/QCamera2/stack/common/cam_types.h
@@ -1673,7 +1673,7 @@ typedef struct {
     int32_t est_snap_iso_value;
     uint32_t est_snap_luma;
     uint32_t est_snap_target;
-    uint32_t xiaomi_reversed;
+    float xiaomi_reversed;
 } cam_3a_params_t;
 
 typedef struct {
